move repository to entities as inner static class
