+++ backend/src/main/java/org/cryptomator/hub/entities/Device.java
@@ -1,7 +1,8 @@
 package org.cryptomator.hub.entities;
 
-import io.quarkus.hibernate.orm.panache.PanacheEntityBase;
+import io.quarkus.hibernate.orm.panache.PanacheRepositoryBase;
 import io.quarkus.panache.common.Parameters;
+import jakarta.enterprise.context.ApplicationScoped;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.EnumType;
@@ -149,4 +150,19 @@ public class Device {
 		return Objects.hash(id, owner, name, type, publickey, userPrivateKey, creationTime);
 	}
 
+	@ApplicationScoped
+	public static class Repository implements PanacheRepositoryBase<Device, String> {
+
+		public Device findByIdAndUser(String deviceId, String userId) throws NoResultException {
+			return find("#Device.findByIdAndOwner", Parameters.with("deviceId", deviceId).and("userId", userId)).singleResult();
+		}
+
+		public Stream<Device> findAllInList(List<String> ids) {
+			return find("#Device.allInList", Parameters.with("ids", ids)).stream();
+		}
+
+		public void deleteByOwner(String userId) {
+			delete("#Device.deleteByOwner", Parameters.with("userId", userId));
+		}
+	}
 }
