+++ backend/src/main/java/org/cryptomator/hub/entities/EffectiveVaultAccess.java
@@ -1,5 +1,8 @@
 package org.cryptomator.hub.entities;
 
+import io.quarkus.hibernate.orm.panache.PanacheRepositoryBase;
+import io.quarkus.panache.common.Parameters;
+import jakarta.enterprise.context.ApplicationScoped;
 import jakarta.persistence.Column;
 import jakarta.persistence.Embeddable;
 import jakarta.persistence.EmbeddedId;
@@ -11,8 +14,11 @@ import jakarta.persistence.Table;
 import org.hibernate.annotations.Immutable;
 
 import java.io.Serializable;
+import java.util.Collection;
+import java.util.List;
 import java.util.Objects;
 import java.util.UUID;
+import java.util.stream.Collectors;
 
 @Entity
 @Immutable
@@ -139,4 +145,33 @@ public class EffectiveVaultAccess {
 		}
 	}
 
+	@ApplicationScoped
+	public static class Repository implements PanacheRepositoryBase<EffectiveVaultAccess, Id> {
+
+		public boolean isUserOccupyingSeat(String userId) {
+			return count("#EffectiveVaultAccess.countSeatsOccupiedBySingleUser", Parameters.with("userId", userId)) > 0;
+		}
+
+		public long countSeatsOccupiedByUsers(List<String> userIds) {
+			return count("#EffectiveVaultAccess.countSeatsOccupiedByUsers", Parameters.with("userIds", userIds));
+		}
+
+		public long countSeatOccupyingUsers() {
+			return count("#EffectiveVaultAccess.countSeatOccupyingUsers");
+		}
+
+		public long countSeatOccupyingUsersWithAccessToken() {
+			return count("#EffectiveVaultAccess.countSeatOccupyingUsersWithAccessToken");
+		}
+
+		public long countSeatOccupyingUsersOfGroup(String groupId) {
+			return count("#EffectiveVaultAccess.countSeatOccupyingUsersOfGroup", Parameters.with("groupId", groupId));
+		}
+
+		public Collection<VaultAccess.Role> listRoles(UUID vaultId, String authorityId) {
+			return find("#EffectiveVaultAccess.findByAuthorityAndVault", Parameters.with("vaultId", vaultId).and("authorityId", authorityId)).stream()
+					.map(eva -> eva.getId().getRole())
+					.collect(Collectors.toUnmodifiableSet());
+		}
+	}
 }
