--- backend/src/main/java/org/cryptomator/hub/entities/GroupRepository.java
@@ -1,8 +1,0 @@
-package org.cryptomator.hub.entities;
-
-import io.quarkus.hibernate.orm.panache.PanacheRepositoryBase;
-import jakarta.enterprise.context.ApplicationScoped;
-
-@ApplicationScoped
-public class GroupRepository implements PanacheRepositoryBase<Group, String> {
-}
