+++ backend/src/main/java/org/cryptomator/hub/entities/events/VaultAccessGrantedEvent.java
@@ -1,5 +1,7 @@
 package org.cryptomator.hub.entities.events;
 
+import io.quarkus.hibernate.orm.panache.PanacheRepository;
+import jakarta.enterprise.context.ApplicationScoped;
 import jakarta.persistence.Column;
 import jakarta.persistence.DiscriminatorValue;
 import jakarta.persistence.Entity;
@@ -65,4 +67,17 @@ public class VaultAccessGrantedEvent extends AuditEvent {
 		return Objects.hash(id, grantedBy, vaultId, authorityId);
 	}
 
+	@ApplicationScoped
+	public static class Repository implements PanacheRepository<VaultAccessGrantedEvent> {
+
+		public void log(String grantedBy, UUID vaultId, String authorityId) {
+			var event = new VaultAccessGrantedEvent();
+			event.setTimestamp(Instant.now());
+			event.setGrantedBy(grantedBy);
+			event.setVaultId(vaultId);
+			event.setAuthorityId(authorityId);
+			persist(event);
+		}
+
+	}
 }
